requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
